+++ CHANGELOG.md
@@ -35,6 +35,7 @@ Changelog for NeoFS Node
 - Policer iterates engine-level object list now instead of shard-level (#3862)
 - SN now ignores `copies_number` field of `object.PutRequest.Body.Init` message (#3830)
 - Policer starts from a random offset (#3879)
+- Every object with non-zero payload is now paid, not only regular ones (#3856)
 
 ### Removed
 - `node.persistent_sessions.path` config option from SN config (#3846)